[proof] Cleanup some terminology around Frozen subtree iterators.
We should be counting in leaves, not in "frozen subtree size".
Also, fix up u64 -> usize to match the previous diff.

+++ types/src/proof/accumulator/accumulator_test.rs
@@ -123,7 +123,7 @@ proptest! {
         let position_to_hash = compute_hashes_for_all_positions(&all_hashes);
 
         let subtree_hashes: Vec<_> =
-            FrozenSubtreeSiblingIterator::new(hashes1.len() as u64, all_hashes.len() as u64)
+            FrozenSubtreeSiblingIterator::new(hashes1.len(), all_hashes.len())
                 .filter_map(|pos| position_to_hash.get(&pos).cloned())
                 .collect();
         let new_accumulator = accumulator
